hero2qlte: Update to G9350ZCU2BQC1 blob

+++ rootdir/etc/init.qcom.sh
@@ -219,7 +219,7 @@ case "$target" in
                   ;;
         esac
         ;;
-    "msm8994" | "msm8992" | "msm8998")
+    "msm8994" | "msm8992" | "msmcobalt")
         start_msm_irqbalance
         ;;
     "msm8996")
@@ -315,28 +315,24 @@ echo 1 > /data/misc/radio/db_check_done
 #
 # Make modem config folder and copy firmware config to that folder for RIL
 #
-#ifdef VENDOR_EDIT
-# modify the source path to /system/etc/firmware/mbn_ota/ , hanqingpu, 20151119
-##if [ -f /data/misc/radio/ver_info.txt ]; then
-##    prev_version_info=`cat /data/misc/radio/ver_info.txt`
-##else
-##    prev_version_info=""
-##fi
+if [ -f /data/misc/radio/ver_info.txt ]; then
+    prev_version_info=`cat /data/misc/radio/ver_info.txt`
+else
+    prev_version_info=""
+fi
 
-##cur_version_info=`cat /firmware/verinfo/ver_info.txt`
-##if [ ! -f /firmware/verinfo/ver_info.txt -o "$prev_version_info" != "$cur_version_info" ]; then
-rm -rf /data/misc/radio/modem_config
-mkdir /data/misc/radio/modem_config
-chmod 770 /data/misc/radio/modem_config
-##    cp -r /firmware/image/modem_pr/mcfg/configs/* /data/misc/radio/modem_config
-cp -r /system/etc/firmware/mbn_ota/* /data/misc/radio/modem_config
-chown -hR radio.radio /data/misc/radio/modem_config
-##    cp /firmware/verinfo/ver_info.txt /data/misc/radio/ver_info.txt
-##    chown radio.radio /data/misc/radio/ver_info.txt
-##fi
-##cp /firmware/image/modem_pr/mbn_ota.txt /data/misc/radio/modem_config
-##chown radio.radio /data/misc/radio/modem_config/mbn_ota.txt
-#endif /*VENDOR_EDIT*/
+cur_version_info=`cat /firmware/verinfo/ver_info.txt`
+if [ ! -f /firmware/verinfo/ver_info.txt -o "$prev_version_info" != "$cur_version_info" ]; then
+    rm -rf /data/misc/radio/modem_config
+    mkdir /data/misc/radio/modem_config
+    chmod 770 /data/misc/radio/modem_config
+    cp -r /firmware/image/modem_pr/mcfg/configs/* /data/misc/radio/modem_config
+    chown -hR radio.radio /data/misc/radio/modem_config
+    cp /firmware/verinfo/ver_info.txt /data/misc/radio/ver_info.txt
+    chown radio.radio /data/misc/radio/ver_info.txt
+fi
+cp /firmware/image/modem_pr/mbn_ota.txt /data/misc/radio/modem_config
+chown radio.radio /data/misc/radio/modem_config/mbn_ota.txt
 echo 1 > /data/misc/radio/copy_complete
 
 #check build variant for printk logging
